+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 Breaking changes:
 
 - rename `:sanitize_locale_by` option `:cast_locale_by`
+- rename `PlugLocale.WebBrowser.build_localized_path` to `PlugLocale.WebBrowser.build_locale_path`
+
+New Features:
+
+- add `PlugLocale.WebBrowser.put_locale_resp_cookie/2` / `PlugLocale.WebBrowser.put_locale_resp_cookie/3`
 
 ## v0.3.0